# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API",
        "description": "API completa para el sistema de ecommerce con autenticación, productos, carrito, pedidos, pagos, cupones y más.",
        "contact": {
            "name": "Soporte Técnico",
            "email": "soporte@ecommerce.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/cart": {
            "get": {
                "tags": [
                    "Carrito"
                ],
                "summary": "Obtener carrito del usuario",
                "description": "Retorna el carrito de compras del usuario autenticado con todos sus items",
                "operationId": "299543e2834f0f740d48dd4129cfb4e6",
                "responses": {
                    "200": {
                        "description": "Carrito obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Carrito obtenido correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "format": "uuid"
                                                            },
                                                            "product": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "string",
                                                                        "format": "uuid"
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "iPhone 15 Pro"
                                                                    },
                                                                    "slug": {
                                                                        "type": "string",
                                                                        "example": "iphone-15-pro"
                                                                    },
                                                                    "image": {
                                                                        "type": "string",
                                                                        "example": "https://example.com/image.jpg"
                                                                    },
                                                                    "category": {
                                                                        "type": "string",
                                                                        "example": "Electrónicos"
                                                                    },
                                                                    "brand": {
                                                                        "type": "string",
                                                                        "example": "Apple"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 999.99
                                                            },
                                                            "original_price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 1099.99
                                                            },
                                                            "subtotal": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 1999.98
                                                            },
                                                            "savings": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 200
                                                            },
                                                            "has_discount": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "discount_percentage": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 18.18
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total_items": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 4999.95
                                                },
                                                "total_savings": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 500
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "title": "Respuesta de Error",
                "description": "Estructura estándar para respuestas de error",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error en la operación"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}